Optimize FieldType in Manager
